requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
google-generativeai==0.8.3
python-decouple==3.8